+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # CHANGELOG
 
-## master
+## 11.1.2 Mar 25, 2023
 
 Changes: